% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concurrent.R
\name{concurrent}
\alias{concurrent}
\title{CONCURRENT MODEL FOR REGRESSION OVER FUNCTIONAL DATA}
\usage{
concurrent()
}
\value{
A training and a prediction function.
}
\description{
This model is a concurrent model, which may be fed to split.conf.fun.
}
\details{
For more details about the structure of the inputs go to split.R
}
