% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean.R
\name{mean_lists}
\alias{mean_lists}
\title{MEAN OF FUNCTIONAL DATA}
\usage{
mean_lists()
}
\value{
It outputs a training function and a prediction function.
}
\description{
This model is a fed to a Functional Conformal Prediction function.
}
\details{
For more details about the structure of the inputs go to split.fun.R
}
